// src/api/generated/models/TaskResponseDto.ts
export type TaskResponseDto = {
    /**
     * @description Task ID
     * @type string
    */
    id: string;
    /**
     * @description External ID
     * @type string
    */
    externalId: string;
    /**
     * @description Task title
     * @type string
    */
    title: string;
    /**
     * @description Task note
     * @type string
    */
    note: string;
    /**
     * @description Task priority
     * @type number
    */
    priority: number;
    /**
     * @description Recurring task
     * @type object | undefined
    */
    recurrence?: object;
    /**
     * @description Journal date
     * @type string
    */
    journalDate: string;
    /**
     * @description Completion status
     * @type number
    */
    complete: number;
    /**
     * @description Last completion date
     * @type string
    */
    completeLast: string;
    /**
     * @description Task state
     * @type number
    */
    state: number;
    /**
     * @description Check status
     * @type number
    */
    checked: number;
    /**
     * @description Show in basket
     * @type boolean
    */
    showInBasket: boolean;
    /**
     * @description Deletion date
     * @type string
    */
    deleteDate: string;
    /**
     * @description Project ID
     * @type string
    */
    projectId: string;
    /**
     * @description Recurrence generator ID
     * @type string
    */
    recurrenceGeneratorId: string;
    /**
     * @description Start date
     * @type string
    */
    start: string;
    /**
     * @description Start notification read
     * @type boolean
    */
    startNotifyReaded: boolean;
    /**
     * @description Notification time
     * @type number
    */
    notify: number;
    /**
     * @description List of read start notifications
     * @type array
    */
    startNotifiesReaded: number[];
    /**
     * @description List of notifications
     * @type array
    */
    notifies: number[];
    /**
     * @description Use time
     * @type boolean
    */
    useTime: boolean;
    /**
     * @description Encrypted content
     * @type string
    */
    crypted: string;
    /**
     * @description Deferred
     * @type boolean
    */
    deferred: boolean;
    /**
     * @description Deadline
     * @type string
    */
    deadline: string;
    /**
     * @description Deadline notification read
     * @type boolean
    */
    deadlineNotifyReaded: boolean;
    /**
     * @description Parent task ID
     * @type string
    */
    parent: string;
    /**
     * @description Group ID
     * @type string
    */
    group: string;
    /**
     * @description Creation date
     * @type string
    */
    createdDate: string;
    /**
     * @description Modification date
     * @type string
    */
    modificatedDate: string;
    /**
     * @description Schedule order
     * @type number
    */
    scheduleOrder: number;
    /**
     * @description Order in parent task
     * @type number
    */
    parentOrder: number;
    /**
     * @description Integration item ID
     * @type string
    */
    integrationItemId: string;
    /**
     * @description Seen today
     * @type string
    */
    seenToday: string;
    /**
     * @description Duration in minutes
     * @type number
    */
    timeLength: number;
    /**
     * @description Pomodoro count
     * @type number
    */
    pomodoroCount: number;
    /**
     * @description Total pomodoro time
     * @type number
    */
    pomodoroTotalTime: number;
    /**
     * @description Alarm notification
     * @type boolean
    */
    alarmNotify: boolean;
    /**
     * @description Is note
     * @type boolean
    */
    isNote: boolean;
    /**
     * @description Tags
     * @type array
    */
    tags: string[];
    /**
     * @description Removed
     * @type boolean
    */
    removed: boolean;
    /**
     * @description Task modifications
     * @type object
    */
    modificated: object;
};